Move InitializeSpec function into appropriate package
