+++ v2/tools/generator/internal/functions/testdata/TestGolden_InitializeSpecFunction_GeneratesExpectedCode/InitializeSpecFunction.golden
@@ -1,0 +1,39 @@
+// Code generated by azure-service-operator-codegen. DO NOT EDIT.
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+package v20200101
+
+import (
+	"fmt"
+	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// +kubebuilder:object:root=true
+// +kubebuilder:subresource:status
+type Person struct {
+	metav1.TypeMeta   `json:",inline"`
+	metav1.ObjectMeta `json:"metadata,omitempty"`
+	Spec              Person_Spec   `json:"spec,omitempty"`
+	Status            Person_STATUS `json:"status,omitempty"`
+}
+
+// InitializeSpec initializes the spec for this resource from the given status
+func (person *Person) InitializeSpec(status genruntime.ConvertibleStatus) error {
+	if s, ok := status.(*Person_STATUS); ok {
+		return person.Spec.Initialize_From_Status(s)
+	}
+
+	return fmt.Errorf("expected Status of type Person_STATUS but received %T instead", status)
+}
+
+// +kubebuilder:object:root=true
+type PersonList struct {
+	metav1.TypeMeta `json:",inline"`
+	metav1.ListMeta `json:"metadata,omitempty"`
+	Items           []Person `json:"items"`
+}
+
+func init() {
+	SchemeBuilder.Register(&Person{}, &PersonList{})
+}
